+++ design-system/docs/playroom/frame.tsx
@@ -2,7 +2,7 @@ import { ReactNode } from 'react';
 
 import {
   ClientSideOnlyDocumentElement,
-  VoussoirProvider,
+  KeystarProvider,
 } from '@keystar/ui/core';
 
 export default function FrameComponent({
@@ -11,13 +11,13 @@ export default function FrameComponent({
   children: ReactNode;
 }): JSX.Element {
   return (
-    <VoussoirProvider>
+    <KeystarProvider>
       <ClientSideOnlyDocumentElement />
       <link
         href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap"
         rel="stylesheet"
       />
       {children}
-    </VoussoirProvider>
+    </KeystarProvider>
   );
 }
