+++ design-system/pkg/src/core/context.tsx
@@ -1,20 +1,20 @@
 import { createContext, useContext } from 'react';
-import { VoussoirProviderContext } from './types';
+import { KeystarProviderContext } from './types';
 
 // Context is in a separate file to avoid fast refresh issue where the
 // old provider context values are immediately replaced with the null default.
-export const Context = createContext<VoussoirProviderContext | null>(null);
-Context.displayName = 'VoussoirProviderContext';
+export const Context = createContext<KeystarProviderContext | null>(null);
+Context.displayName = 'KeystarProviderContext';
 
 /**
  * Returns the settings and styles applied by the nearest parent
  * Provider. Properties explicitly set by the nearest parent Provider override
  * those provided by preceeding Providers.
  */
-export function useProvider(): VoussoirProviderContext {
+export function useProvider(): KeystarProviderContext {
   let context = useContext(Context);
   if (!context) {
-    throw new Error('Attempt to access context outside of VoussoirProvider.');
+    throw new Error('Attempt to access context outside of KeystarProvider.');
   }
   return context;
 }
