+++ design-system/pkg/src/core/index.ts
@@ -1,7 +1,7 @@
-export { VoussoirProvider } from './VoussoirProvider';
+export { KeystarProvider } from './KeystarProvider';
 export { useProvider, useProviderProps } from './context';
 export { ClientSideOnlyDocumentElement } from './ClientSideOnlyDocumentElement';
 export { documentElementClasses } from './globals';
 export { TestProvider } from './TestProvider';
 
-export type { VoussoirProviderProps } from './types';
+export type { KeystarProviderProps } from './types';
