+++ design-system/pkg/src/list-view/ListViewItem.tsx
@@ -19,7 +19,7 @@ import { assert } from 'emery';
 import React, { useRef } from 'react';
 
 import { Checkbox } from '@keystar/ui/checkbox';
-import { VoussoirProvider } from '@keystar/ui/core';
+import { KeystarProvider } from '@keystar/ui/core';
 import { Icon } from '@keystar/ui/icon';
 import { chevronLeftIcon } from '@keystar/ui/icon/icons/chevronLeftIcon';
 import { chevronRightIcon } from '@keystar/ui/icon/icons/chevronRightIcon';
@@ -215,7 +215,7 @@ export function ListViewItem<T>(props: ListViewItemProps<T>) {
     item.rendered
   );
   if (isDisabled) {
-    content = <VoussoirProvider isDisabled>{content}</VoussoirProvider>;
+    content = <KeystarProvider isDisabled>{content}</KeystarProvider>;
   }
 
   return (
